Fix build break
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@56391 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ domainjoin/libdomainjoin/src/djkrb5conf.c
@@ -1298,7 +1298,7 @@ DJModifyKrb5Conf(
         {
             GCE(ceError = ERROR_INVALID_PARAMETER);
         }
-        bAssumeDefaultDomain = (DWORD)pReadData;
+        bAssumeDefaultDomain = (DWORD)(size_t)pReadData;
     }
     else
     {
